requirements.txt
flask==3.1.0
flask-cors==5.0.1
firebase-admin==6.6.0
python-dotenv==1.1.0
werkzeug==3.1.3
gunicorn==23.0.0
simple-salesforce==1.12.6
